Add tests for all HSE params
Parameters are part of the public API. We need to make sure that changes
to parameters are understood when they break tests or how values are
parsed.

+++ lib/mpool/include/mpool/mpool_structs.h
@@ -47,10 +47,10 @@ enum mpool_mclass {
  */
 struct mpool_cparams {
     struct {
-        size_t  fmaxsz;
-        size_t  mblocksz;
-        uint8_t filecnt;
-        char    path[PATH_MAX];
+        uint64_t fmaxsz;
+        uint32_t mblocksz;
+        uint8_t  filecnt;
+        char     path[PATH_MAX];
     } mclass[MP_MED_COUNT];
 };
